+++ traktor/engine/config_mixin.py
@@ -12,11 +12,11 @@ class ConfigMixin:
     def config_list() -> List[ConfigEntry]:
         """List all configuration values."""
         return [
-            ConfigEntry(key=ConfigEntry.Key.format, value=config.format.value),
-            ConfigEntry(key=ConfigEntry.Key.db_path, value=config.db_path),
-            ConfigEntry(
-                key=ConfigEntry.Key.timezone, value=config.timezone.zone
-            ),
+            ConfigEntry(key=ConfigKey.format, value=config.format.value),
+            ConfigEntry(key=ConfigKey.timezone, value=config.timezone.zone),
+            ConfigEntry(key=ConfigKey.prod_db_path, value=config.prod_db_path),
+            ConfigEntry(key=ConfigKey.test_db_path, value=config.test_db_path),
+            ConfigEntry(key=ConfigKey.use_test_db, value=config.use_test_db),
         ]
 
     @classmethod
@@ -29,14 +29,17 @@ class ConfigMixin:
                 raise errors.InvalidConfiguration(
                     value=value, valid_values=valid_values
                 )
-
-        elif key == ConfigKey.db_path:
-            config.db_path = value
         elif key == ConfigKey.timezone:
             try:
                 config.timezone = pytz.timezone(value)
             except Exception:
                 raise errors.InvalidConfiguration(value=value)
+        elif key == ConfigKey.prod_db_path:
+            config.prod_db_path = value
+        elif key == ConfigKey.test_db_path:
+            config.test_db_path = value
+        elif key == ConfigKey.use_test_db:
+            config.use_test_db = value.lower() in ("true", "on")
 
         config.save()